azures-storage/function-app/requirements.txt
azure-functions
azure-communication-email
azure-storage-blob
